requirements.txt
asgiref==3.6.0
cloudinary==1.31.0
crispy-bootstrap5==0.7
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.16
django-choices==1.7.2
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.2
gunicorn==20.1.0
phonenumbers==8.13.5
Pillow==9.4.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
